+++ src/contracts/addresses.ts
@@ -1,9 +1,9 @@
 const FUJI_ADDRESSES = {
     "UserManager": "0x36BdBF70B474D6A2aF3F17531Cfe1a9Ea1036c03",
-    "FriendKeyManager": "0x959AF465b9C8d34BD37c9e61b8d457daAEEac241",
-    "FriendKey0": "0x808bbdF7CB4e4Cd2DA518515B13a8bc642367ed7",
-    "FriendKey1": "0xe4806d9ecD28c2e54F7c913698267Ea8F828765A",
-    "FriendKey2": "0x73af150C6dDc8f27E2Af0cc55fcAC32f7A72Dc53",
+    "FriendKeyManager": "0xb64aFd1277EF94F214e4BF8F51B7018Fa0F7C9c9",
+    "FriendKey0": "0xC4Dd875618429b8E015f5B612Ed4d18af54d34E7",
+    "FriendKey1": "0xd8B16F1007A04a12b84440B96a055f1788BB3d43",
+    "FriendKey2": "0x5ff752925cC6223661d18F384940D4223FE5F92b",
 }
