+++ src/pages/_app.tsx
@@ -1,11 +1,8 @@
-import MainLayout from "@/components/ui/Layout/MainLayout";
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
-    <MainLayout>
-      <Component {...pageProps} />
-    </MainLayout>
+    <Component {...pageProps} />
   );
 }
