+++ src/pages/index.tsx
@@ -1,3 +1,5 @@
+import MainLayout from "@/components/ui/Layout/MainLayout";
+
 export default function Home() {
-  return <main></main>;
+  return <MainLayout>Index</MainLayout>
 }
